backend/requirements/base.txt
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
black==25.1.0
click==8.2.1
Django==5.2.2
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
django-environ==0.12.0
djangorestframework==3.16.0
drf-spectacular==0.28.0
flake8==7.2.0
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pycodestyle==2.13.0
pyflakes==3.3.2
pyotp==2.9.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.14.0
uritemplate==4.2.0
